==> Lessons/06_Lab_GLMs.Rmd <==
---
title: "6: Part 1 - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2022"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: inline
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Additional comments on videos - t-test
4. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
library(agricolae)   

PeterPaul.chem.nutrients <- read.csv("../Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

EPAair <- read.csv("../Data/Processed/EPAair_O3_PM25_NC1819_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")


Litter <- read.csv("../Data/Processed/NEON_NIWO_Litter_mass_trap_Processed.csv", stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```

## T-Test

Continuous response, one categorical explanatory variable with two categories (or comparison to a single value if a one-sample test).

### Formulating Hypothesis for µ

Two hypotheses are formed – the null hypothesis and the alternative hypothesis.
The null hypothesis and the alternative hypothesis combine to cover all possible values for the population mean.
The null hypothesis must have the equality.
The null and alternative hypotheses are always stated in terms of the population mean (mu).

### One-sample t-test
The object of a one sample test is to test the null hypothesis that the mean of the group is equal to a specific value. For example, we might ask ourselves (from the EPA air quality processed dataset): 

Function t.test()
**x**	a (non-empty) numeric vector of data values.
**alternative** a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less". You can specify just the initial letter.
**mu** a number indicating the true value of the mean (or difference in means if you are performing a two sample test).
**formula** a formula of the form lhs ~ rhs where lhs is a numeric variable giving the data values and rhs either 1 for a one-sample or paired test or a factor with two levels giving the corresponding groups. If lhs is of class "Pair" and rhs is 1, a paired test is done.


Are Ozone levels below the threshold for "good" AQI index (0-50)?


> Exercise 1: State the hypotheses for testing mean of AQI index.

> Answer: H0: AQI mu <= "good" (50)
H1: AQI mu > 50 


```{r}


summary(EPAair$Ozone)
EPAair.subsample <- sample_n(EPAair, 5000) #sampling 5000 data points

# Evaluate assumption of normal distribution
shapiro.test((EPAair.subsample$Ozone)) #shapiro test null = the data is normal
ggplot(EPAair, aes(x = Ozone)) +
  geom_histogram() 
qqnorm(EPAair$Ozone); qqline(EPAair$Ozone)

O3.onesample <- t.test(EPAair$Ozone, mu = 50, alternative = "less") #she actualy has the null = mean greater than 50
O3.onesample #the p value shows that the alternative hypothesis is statisticall significant

Ozone.plot <- ggplot(EPAair, aes(x = Ozone)) +
  #geom_density(stat = "count", fill = "gray") +
  geom_density(fill = "gray") +
  geom_vline(xintercept = 50, color = "#238b45", lty = 2, size = 0.9) +
  scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0))
print(Ozone.plot)
```

Write a sentence or two about the results of this test. Include both the results of the test and an interpretation that puts the findings in context of the research question.

> Results: we reject the null that it's above 50. We're confident that we've got good AQI

### Two-sample t-test

The two-sample *t* test is used to test the hypothesis that the mean of two samples is equivalent. Unlike the one-sample tests, a two-sample test requires a second assumption that the variance of the two groups is equivalent. Are Ozone levels different between 2018 and 2019?

```{r}
shapiro.test(EPAair$Ozone[EPAair$Year == 2018]) #not normal
shapiro.test(EPAair$Ozone[EPAair$Year == 2019]) #not normal

#p-value less than 0.05 then reject null for 2018 and 2019 i.e. data do not follow normal distribution

#Compare variance using F-test (only)
var.test(EPAair$Ozone ~ EPAair$Year) #reject the null that variance ratio = 1

#p-value less than 0.05 then reject null for 2018 and 2019 i.e. true ratio not equal to one

ggplot(EPAair, aes(x = Ozone, color = as.factor(Year))) +
  geom_freqpoly()

# Format as a t-test
O3.twosample <- t.test(EPAair$Ozone ~ EPAair$Year) #regressed over two level categorical variable 
O3.twosample
O3.twosample$p.value #we can literally just call the p-value on it's own

# Format as a GLM
O3.twosample2 <- lm(EPAair$Ozone ~ EPAair$Year) #we're trying to predict ozone with year
summary(O3.twosample2)

plot(O3.twosample2)


```


## Statistical Test: Cheat sheet

**F-test:** Compare the variances of two groups. The data must be normally distributed.

**Bartlett’s test:** Compare the variances of two or more groups. The data must be normally distributed.

**Shapiro.test:** check for normality

**One-sample t-test:** check if mean is equal/less/greater to specific value, single variable

**Two-sample t-test:**  check if mean of two samples is equivalent

### Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer). (Ozone = Y X = year) - this is how she phrases it
2. geom_boxplot of ozone divided by year . Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r}
#Exercise 2:
Exercise_2.plot <- ggplot(EPAair, aes(x = Ozone, fill = as.factor(Year))) + #the as.factor() was missing at first, that was what was holding me up - it was originally an integer and we want it as a categorical variable
  geom_density(alpha =0.5) +
  scale_fill_manual(values = c("darkgreen", "darkblue")) +
  labs(x= "Ozone AQI Value", y= "Density", fill = "")
print(Exercise_2.plot)

Exercise_2_box.plot <-
  ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_boxplot() +
  stat_summary(geom = "text", fun.y= max, vjust = -1, size = 4, label = c("b", "a")) +
  labs(x = "", y = "Ozone AQI Value")
print(Exercise_2_box.plot)

#the fun.y puts the label on the max y value for each x variable -> that's what's going on in module 6 part 2fellow fucking off


Exercise_2_violin.plot <-
  ggplot(EPAair, aes(x = as.factor(Year), y = Ozone)) +
  geom_violin(draw_quantiles = 0.5) +
  stat_summary(geom = "text" , fun.y = max, vjust = -1, size = 4, label = c("b", "a")) +
  labs(x = "", y = "Ozone AQI Value")
print(Exercise_2_violin.plot)


```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

```{r}

#Exercise 3: Run a linear regression PM2.5(y) by Ozone(x). Find the p-value and R-squared value. 
Exercise_3_lm <- lm(data = EPAair, PM2.5 ~ Ozone) #we're trying to predict ozone with year
summary(Exercise_3_lm)

plot(Exercise_3_lm)

#p-value: basically 0
#R-squared: 0.1316

#Exercise 4: Build a scatterplot. Add a line and standard error for the linear regression. 
Exercise_3_plot <- ggplot(EPAair, aes(x = Ozone, y = PM2.5)) +
  geom_point() +
  geom_smooth(method = lm)
print(Exercise_3_plot)


```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 5: Build correlation plots and identify more possible explanatory variables to add to the regression.

library(corrplot)
EPAair_working <- 
  EPAair %>%
  select(Year, Month, meanLat:Ozone) %>%
  na.omit()

EPAairCorr <- cor(EPAair_working)
corrplot(EPAairCorr, method = "ellipse")
corrplot.mixed(EPAairCorr, upper = "ellipse") #one side shoes the values, the other the ellipse



#Exercise 6: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 5?

TPAIC <- lm(data = EPAair, PM2.5 ~ Ozone + Year + Month +
              meanLat + meanLng)
step(TPAIC)
#Apparently including, Month, meanLat, meanLng is important

#No I was not going to select some of those


#Exercise 7: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 5. 

Exercise_7 <-lm(data = EPAair, PM2.5 ~ Ozone + Month +
              meanLat + meanLng)
summary(Exercise_7)

#R-squared = 19.26% compared to the single lm which had an R-squared fo 13.14%
#Apparently these 3 additional variables are crucial for understanding stuff
```
